moved common profile fns into profiler utils
